chore: use new artwork screen sections
