+++ src/app/Scenes/HomeView/Sections/ArtworksRailHomeViewSection.tsx
@@ -22,7 +22,7 @@ export const ArtworksRailHomeViewSection: React.FC<ArtworksRailHomeViewSectionPr
   const data = useFragment(fragment, section)
   const title = data.component?.title
   const artworks = extractNodes(data.artworksConnection)
-  const componentHref = `/home-view/section/${data.internalID}`
+  const componentHref = data.component?.behaviors?.viewAll?.href
 
   if (!artworks || artworks.length === 0) {
     return null
@@ -51,6 +51,7 @@ export const ArtworksRailHomeViewSection: React.FC<ArtworksRailHomeViewSectionPr
             onPress={
               componentHref
                 ? () => {
+                    console.log({ componentHref })
                     navigate(componentHref)
                   }
                 : undefined
