+++ src/app/Scenes/HomeViewSectionScreenArtworks/HomeViewSectionScreenArtworks.tsx
@@ -4,7 +4,8 @@ import { HomeViewSectionScreenArtworksQuery } from "__generated__/HomeViewSectio
 import { HomeViewSectionScreenArtworks_artworksRailHomeViewSection$key } from "__generated__/HomeViewSectionScreenArtworks_artworksRailHomeViewSection.graphql"
 import { MasonryInfiniteScrollArtworkGrid } from "app/Components/ArtworkGrids/MasonryInfiniteScrollArtworkGrid"
 import { PAGE_SIZE } from "app/Components/constants"
-import { HomeViewSectionScreenArtworksPlaceholder } from "app/Scenes/HomeViewSectionScreen/Artworks/HomeViewSectionScreenArtworksPlaceholder"
+import { HomeViewSectionScreenArtworksPlaceholder } from "app/Scenes/HomeViewSectionScreenArtworks/HomeViewSectionScreenArtworksPlaceholder"
+import { goBack } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
 import { withSuspense } from "app/utils/hooks/withSuspense"
 import { NUM_COLUMNS_MASONRY } from "app/utils/masonryHelpers"
@@ -17,7 +18,6 @@ interface ArtworksScreenHomeSection {
 }
 
 export const HomeViewSectionScreenArtworks: React.FC<ArtworksScreenHomeSection> = (props) => {
-  const { scrollHandler } = Screen.useListenForScreenScroll()
   const { data, isLoadingNext, loadNext, refetch, hasNext } = usePaginationFragment<
     HomeViewSectionScreenArtworksQuery,
     HomeViewSectionScreenArtworks_artworksRailHomeViewSection$key
@@ -27,37 +27,43 @@ export const HomeViewSectionScreenArtworks: React.FC<ArtworksScreenHomeSection> 
 
   const RefreshControl = useRefreshControl(refetch)
 
+  const { scrollHandler } = Screen.useListenForScreenScroll()
+
   return (
-    <Flex style={{ height: "100%" }}>
-      <MasonryInfiniteScrollArtworkGrid
-        animated
-        artworks={artworks}
-        numColumns={NUM_COLUMNS_MASONRY}
-        disableAutoLayout
-        pageSize={PAGE_SIZE}
-        ListEmptyComponent={
-          <SimpleMessage m={2}>Nothing yet. Please check back later.</SimpleMessage>
-        }
-        ListHeaderComponent={() => (
-          <Flex>
-            <Text variant="lg-display">{data.component?.title}</Text>
-            <Text variant="xs" pt={2}>
-              {data.artworksConnection?.totalCount} {pluralize("Artwork", artworks.length)}
-            </Text>
-          </Flex>
-        )}
-        refreshControl={RefreshControl}
-        hasMore={hasNext}
-        loadMore={() => {
-          loadNext(PAGE_SIZE)
-        }}
-        isLoading={isLoadingNext}
-        onScroll={scrollHandler}
-        style={{ paddingBottom: 120 }}
-        contextModule={data.internalID as ContextModule}
-        contextScreenOwnerType={data.internalID as ScreenOwnerType}
-      />
-    </Flex>
+    <>
+      <Screen.AnimatedHeader onBack={goBack} title={data.component?.title || ""} />
+
+      <Screen.Body fullwidth>
+        <MasonryInfiniteScrollArtworkGrid
+          animated
+          artworks={artworks}
+          numColumns={NUM_COLUMNS_MASONRY}
+          disableAutoLayout
+          pageSize={PAGE_SIZE}
+          ListEmptyComponent={
+            <SimpleMessage m={2}>Nothing yet. Please check back later.</SimpleMessage>
+          }
+          ListHeaderComponent={() => (
+            <Flex>
+              <Text variant="lg-display">{data.component?.title}</Text>
+              <Text variant="xs" pt={2}>
+                {data.artworksConnection?.totalCount} {pluralize("Artwork", artworks.length)}
+              </Text>
+            </Flex>
+          )}
+          refreshControl={RefreshControl}
+          hasMore={hasNext}
+          loadMore={() => {
+            loadNext(PAGE_SIZE)
+          }}
+          isLoading={isLoadingNext}
+          onScroll={scrollHandler}
+          style={{ paddingBottom: 120 }}
+          contextModule={data.internalID as ContextModule}
+          contextScreenOwnerType={data.internalID as ScreenOwnerType}
+        />
+      </Screen.Body>
+    </>
   )
 }
 
@@ -98,13 +104,13 @@ export const artworksQuery = graphql`
   }
 `
 interface ArtworksScreenHomeSectionQRProps {
-  sectionId: string
+  sectionID: string
 }
 
 export const HomeViewSectionScreenArtworksQueryRenderer: React.FC<ArtworksScreenHomeSectionQRProps> =
   withSuspense((props) => {
     const data = useLazyLoadQuery<HomeViewSectionScreenArtworksQuery>(artworksQuery, {
-      id: props.sectionId,
+      id: props.sectionID,
     })
 
     // This won't happen because the query would fail thanks to the @principalField
@@ -113,5 +119,9 @@ export const HomeViewSectionScreenArtworksQueryRenderer: React.FC<ArtworksScreen
       return <Text>Something went wrong.</Text>
     }
 
-    return <HomeViewSectionScreenArtworks section={data.homeView.section} />
+    return (
+      <Screen>
+        <HomeViewSectionScreenArtworks section={data.homeView.section} />
+      </Screen>
+    )
   }, HomeViewSectionScreenArtworksPlaceholder)
